requirements.txt
celery==4.2.0
certifi==2018.8.24
coreapi==2.3.3
Django==2.0.5
django-debug-toolbar==1.9.1
django-model-utils==3.1.2
django-rest-framework==0.1.0
django-simple-history==2.4.0
djangorestframework==3.8.2
jsonfield2==3.0.1
Markdown==2.6.11
mysqlclient==1.3.13
numpy==1.15.1
pandas==0.23.4
Pygments==2.2.0
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-Levenshtein==0.12.0
requests==2.19.1
scipy==1.1.0
